requirements.txt
# Core dependencies
grpcio>=1.54.0
grpcio-tools>=1.54.0
protobuf>=4.23.0
fastapi>=0.95.2
uvicorn>=0.22.0
pydantic>=1.10.7
python-dotenv>=1.0.0

# PDF Processing
PyMuPDF>=1.22.3
pdfplumber>=0.9.0

# ML & Embeddings
openai>=0.27.8
sentence-transformers>=2.2.2
instructor>=0.2.3

# Vector Database
qdrant-client>=1.1.7
weaviate-client>=3.15.4

# Utilities
loguru>=0.7.0
structlog>=23.1.0
pytest>=7.3.1
redis>=4.5.5
prometheus-client>=0.17.0